Relocate some classification functionality
