+++ src/bin/load_classifier_output.rs
@@ -1,12 +1,8 @@
-use biodivine_lib_param_bn::symbolic_async_graph::GraphColors;
-use biodivine_lib_param_bn::BooleanNetwork;
+//! Binary for testing loading of classifier output.
 
-use biodivine_hctl_model_checker::model_checking::get_extended_symbolic_graph;
+use biodivine_hctl_model_checker::bn_classification::load_classifier_output;
 
 use clap::Parser;
-
-use std::fs::{read_dir, read_to_string, File};
-use std::path::PathBuf;
 
 /// Structure to collect CLI arguments
 #[derive(Parser)]
@@ -19,34 +15,14 @@ struct Arguments {
     results_dir: String,
 }
 
+/// Wrapper function to invoke the loading process and feed it with CLI arguments.
 fn main() {
     let args = Arguments::parse();
-    let results_dir = args.results_dir.as_str();
 
-    // load number of HCTL variables from computation metadata
-    let metadata_file_path = PathBuf::from(results_dir).join("metadata.txt");
-    let num_hctl_vars: u16 = read_to_string(metadata_file_path)
-        .unwrap()
-        .parse::<u16>()
-        .unwrap();
+    // load the color sets that represent the classification results
+    let color_sets = load_classifier_output(args.results_dir.as_str(), args.model_path.as_str());
 
-    let model_string = read_to_string(args.model_path).unwrap();
-    let bn = BooleanNetwork::try_from(model_string.as_str()).unwrap();
-    let graph = get_extended_symbolic_graph(&bn, num_hctl_vars);
-
-    let files = read_dir(results_dir).unwrap();
-
-    // expects only BDD dumps (individual files) and a report&metadata in the directory (for now)
-    for file in files {
-        let path = file.unwrap().path().clone();
-        let file_name = path.file_name().unwrap().to_str().unwrap();
-        if file_name == "report.txt" || file_name == "metadata.txt" {
-            continue;
-        }
-        let mut file = File::open(path).unwrap();
-        let bdd = biodivine_lib_bdd::Bdd::read_as_string(&mut file).unwrap();
-
-        let color_set = GraphColors::new(bdd, graph.symbolic_context());
+    for color_set in color_sets {
         println!("{}", color_set.exact_cardinality());
     }
 }
